requirements.txt
asgiref==3.4.1
cycler==0.10.0
DAWG-Python==0.7.2
deskew==0.10.36
Django==3.2.8
django-crispy-forms==1.13.0
docopt==0.6.2
imageio==2.9.0
intervaltree==3.1.0
ipymarkup==0.9.0
kiwisolver==1.3.2
matplotlib==3.4.3
natasha==1.4.0
navec==0.10.0
networkx==2.6.3
numpy==1.21.3
opencv-python==4.5.4.58
pdf2image==1.16.0
Pillow==8.4.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.0
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
razdel==0.5.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
slovnet==0.5.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tifffile==2021.10.12
yargy==0.15.0
